refactor: move http components docs to separate folder
