+++ src/main/docs/components.ts
@@ -1,7 +1,4 @@
-import { badRequest } from './components/bad-request'
-import { forbidden } from './components/forbidden'
-import { serverError } from './components/server-error'
-import { unauthorized } from './components/unauthorized'
+import { badRequest, unauthorized, forbidden, serverError } from '@/main/docs/components/http'
 
 export default {
   badRequest,
