+++ src/main/docs/components/http/bad-request.ts
