+++ src/main/docs/components/http/forbidden.ts
