+++ src/main/docs/components/http/index.ts
@@ -1,0 +1,4 @@
+export * from './bad-request'
+export * from './forbidden'
+export * from './server-error'
+export * from './unauthorized'
